PMM-10612 refactor adding folders
