+++ public/app/percona/shared/components/PerconaBootstrapper/PerconaNavigation/PerconaNavigation.constants.ts
@@ -241,6 +241,11 @@ export const PMM_ENVIRONMENT_OVERVIEW_PAGE: NavModelItem = {
   section: NavSection.Core,
 };
 
+/**
+ * Mapping of menu items id to folders name.
+ *
+ * Folders are created based on the folder name in grafana-dashboards.
+ */
 export const NAV_FOLDER_MAP: Record<string, string> = {
   system: 'OS',
   mysql: 'MySQL',
